Release 0.0.1 alpha
Created a standalone Windows release of what has been made so far. The
game currently allows exploration of one level of a dungeon, fights with
Rats and Radioactive Rats, and several items can be collected including
healing potions and scrolls. If all monsters on the level are defeated
there will be nothing else to do.